Improve DataGrid.stories
